jsonrpc: uds support for Windows
I could not find a way to get mio 0.7 running with Windows' uds. So,
until https://github.com/tokio-rs/mio/issues/880 (https://github.com/deprecrated/mio-uds/issues/7)
is implemented in Mio we use a simple blocking event loop for Windows.

That's fine, it's just for the RPC. We could potentially see some load
on a server, but it'd be on UNIX so Everything Is Fine :tm:.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>